Move mark as completed action to encounter provider level
